improvement: make cipher/mac/kex checks descriptive
When running the current version of tests and a spec fails, we get an output similar to this:

    2) check sshd_config File "/etc/ssh/sshd_config" content should valid cipher
    Failure/Error: should valid_cipher
     expected "..." to valid cipher

This is not very helpful, since it doesn't tell us what it really expects (e.g. a list of valid ciphers). Best case scenario: you only have to look into the spec file, worst case: your spec has some other error (like OS-detection) and it starts to get really confusing.

This implementation fixes this issue. It is not as elegant as the current one, but prints very descriptive errors:

    2) check ssh_config File "/etc/ssh/ssh_config" content should match /^Ciphers aes256-ctr,aes192-ctr,aes128-ctr$/

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>